requirements.txt
streamlit
python-dotenv
boto3
duckdb
pandas
openai
ollama
watchdog